+++ CHANGELOG.md
@@ -21,8 +21,9 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 - Add manual re-rendering of introtext and outrotext
 
 ## 1.3.0
-
-- Add introtext and outrotext rendering for the three main form types
+- Added Pixelletter DocType
+- Added logic for pixelletter sender
+- Added Pixelletter button to Invoice Doctype with dialog
 
 ## 1.2.0